pyproject.toml
[tool.poetry]
name = "gtm-manager-boilerplate"
version = "0.1.0"
description = "Public boilerplate for automating Google Tag Manager workflows."
authors = ["GTM Manager Contributors <opensource@example.com>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.10"
google-api-python-client = "^2.147.0"
google-auth = "^2.34.0"
google-auth-oauthlib = "^1.2.4"
PyYAML = "^6.0.2"

[tool.poetry.group.dev.dependencies]
mypy = "^1.19.1"
pre-commit = "^3.8.0"
ruff = "^0.15.2"
pytest = "^9.0.2"
types-pyyaml = "^6.0.12.20250915"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
strict_optional = true

[tool.ruff]
line-length = 100
target-version = "py310"
extend-select = [
  "I",  # isort-compatible import sorting
  "B",  # bugbear
]

[tool.ruff.lint]
ignore = []
